requirements.txt
google-genai==1.52.0
streamlit==1.51.0
